+++ feature/home/build.gradle.kts
@@ -3,15 +3,10 @@ plugins {
 }
 
 android {
-    namespace = "com.example.main"
+    namespace = "com.example.home"
 }
 
 dependencies {
-    implementation(libs.kotlinx.immutable)
-    implementation(projects.feature.login)
-    implementation(projects.feature.home)
-    implementation(projects.feature.map)
-    implementation(projects.feature.mypage)
     implementation(projects.core.domain)
     implementation(projects.core.data)
     implementation(libs.androidx.core.ktx)
