+++ feature/home/src/main/java/com/example/home/presentation/EventDetailScreen.kt
@@ -1,10 +1,11 @@
-package com.example.main.presentation
+package com.example.home.presentation
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.filled.ArrowBack
+import androidx.compose.material3.Button
 import androidx.compose.material3.CenterAlignedTopAppBar
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
@@ -16,28 +17,29 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
-import com.example.main.R
+import com.example.home.R
 
 @Composable
-fun EditProfileScreen() {
+fun EventDetailScreen() {
     Scaffold(
         modifier = Modifier.fillMaxSize(),
         topBar = {
             CenterAlignedTopAppBar(
-                title = { Text(text = "프로필 편집") },
+                title = {
+                },
                 navigationIcon = {
-                    IconButton(onClick = {}) {
+                    IconButton(onClick = { }) {
                         Icon(
-                            imageVector = Icons.AutoMirrored.Filled.ArrowBack,
-                            contentDescription = "top_bar_nav_edit_profile",
+                            imageVector = Icons.AutoMirrored.Default.ArrowBack,
+                            contentDescription = "top_bar_nav_icon_event_detail",
                         )
                     }
                 },
                 actions = {
-                    IconButton(onClick = {}) {
+                    IconButton(onClick = { }) {
                         Icon(
-                            painter = painterResource(R.drawable.baseline_save_24),
-                            contentDescription = "top_bar_icon_edit_profile_save",
+                            painterResource(R.drawable.baseline_bookmarks_24),
+                            contentDescription = "top_bar_icon_event_detail_bookmark",
                         )
                     }
                 },
@@ -50,13 +52,22 @@ fun EditProfileScreen() {
                 .padding(innerPadding),
             horizontalAlignment = Alignment.CenterHorizontally,
         ) {
-            Text(text = "MyPageScreen", style = MaterialTheme.typography.displayLarge)
+            Text(text = "EventDetailScreen", style = MaterialTheme.typography.displaySmall)
+            Button(onClick = {}) {
+                Text(text = "지도로 보기")
+            }
+            Button(onClick = {}) {
+                Text(text = "전체 참여자 목록 보기")
+            }
+            Button(onClick = {}) {
+                Text(text = "참석 하기")
+            }
         }
     }
 }
 
 @Preview(showBackground = true)
 @Composable
-fun EditProfileScreenPreview() {
-    EditProfileScreen()
+fun EventDetailPreview() {
+    EventDetailScreen()
 }
