+++ feature/home/src/main/java/com/example/home/presentation/HomeScreen.kt
@@ -1,9 +1,13 @@
-package com.example.main.presentation
+package com.example.home.presentation
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Add
 import androidx.compose.material3.Button
+import androidx.compose.material3.FloatingActionButton
+import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
@@ -13,9 +17,17 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 
 @Composable
-fun MapScreen() {
+fun HomeScreen() {
     Scaffold(
         modifier = Modifier.fillMaxSize(),
+        floatingActionButton = {
+            FloatingActionButton(onClick = {}) {
+                Icon(
+                    imageVector = Icons.Default.Add,
+                    contentDescription = "Fab_HomeScreen",
+                )
+            }
+        },
     ) { innerPadding ->
         Column(
             modifier = Modifier
@@ -24,6 +36,10 @@ fun MapScreen() {
             horizontalAlignment = Alignment.CenterHorizontally,
         ) {
             Text(text = "HomeScreen", style = MaterialTheme.typography.displayLarge)
+            // topBar 구현 해야함
+            Button(onClick = {}) {
+                Text(text = "로그인 버튼")
+            }
             Button(onClick = {}) {
                 Text(text = "이벤트 클릭")
             }
@@ -33,6 +49,6 @@ fun MapScreen() {
 
 @Preview(showBackground = true)
 @Composable
-fun MapScreenPreview() {
-    MapScreen()
+fun HomeScreenPreview() {
+    HomeScreen()
 }
