+++ feature/map/src/main/java/com/example/map/presentation/MapScreen.kt
@@ -1,0 +1,38 @@
+package com.example.map.presentation
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Button
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Scaffold
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.tooling.preview.Preview
+
+@Composable
+fun MapScreen() {
+    Scaffold(
+        modifier = Modifier.fillMaxSize(),
+    ) { innerPadding ->
+        Column(
+            modifier = Modifier
+                .fillMaxSize()
+                .padding(innerPadding),
+            horizontalAlignment = Alignment.CenterHorizontally,
+        ) {
+            Text(text = "HomeScreen", style = MaterialTheme.typography.displayLarge)
+            Button(onClick = {}) {
+                Text(text = "이벤트 클릭")
+            }
+        }
+    }
+}
+
+@Preview(showBackground = true)
+@Composable
+fun MapScreenPreview() {
+    MapScreen()
+}
