+++ feature/mypage/src/main/java/com/example/mypage/presentation/EditProfileScreen.kt
@@ -1,11 +1,10 @@
-package com.example.main.presentation
+package com.example.mypage.presentation
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.filled.ArrowBack
-import androidx.compose.material3.Button
 import androidx.compose.material3.CenterAlignedTopAppBar
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
@@ -15,20 +14,30 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
+import com.example.mypage.R
 
 @Composable
-fun SettingScreen() {
+fun EditProfileScreen() {
     Scaffold(
         modifier = Modifier.fillMaxSize(),
         topBar = {
             CenterAlignedTopAppBar(
-                title = { Text(text = "설정") },
+                title = { Text(text = "프로필 편집") },
                 navigationIcon = {
                     IconButton(onClick = {}) {
                         Icon(
                             imageVector = Icons.AutoMirrored.Filled.ArrowBack,
-                            contentDescription = "top_bar_nav_setting",
+                            contentDescription = "top_bar_nav_edit_profile",
+                        )
+                    }
+                },
+                actions = {
+                    IconButton(onClick = {}) {
+                        Icon(
+                            painter = painterResource(R.drawable.baseline_save_24),
+                            contentDescription = "top_bar_icon_edit_profile_save",
                         )
                     }
                 },
@@ -41,19 +50,13 @@ fun SettingScreen() {
                 .padding(innerPadding),
             horizontalAlignment = Alignment.CenterHorizontally,
         ) {
-            Text(text = "SettingScreen", style = MaterialTheme.typography.displayLarge)
-            Button(onClick = {}) {
-                Text(text = "푸시 알림 받기")
-            }
-            Button(onClick = {}) {
-                Text(text = "회원 탈퇴")
-            }
+            Text(text = "MyPageScreen", style = MaterialTheme.typography.displayLarge)
         }
     }
 }
 
 @Preview(showBackground = true)
 @Composable
-fun SettingScreenPreview() {
-    SettingScreen()
+fun EditProfileScreenPreview() {
+    EditProfileScreen()
 }
